finished creating logging page and connected to home page and ajust some codes to make the website more efficient and secure

+++ includes/admin/delete_product.php
@@ -1,5 +1,9 @@
 <?php
-$id = $_GET['id'];
-mysqli_query($conn, "DELETE FROM products WHERE id = $id");
-header("Location: product.php");
+  require_once __DIR__ . '/../auth_functions.php';
+  require_login();
+  require_admin();
+
+  $id = $_GET['id'];
+  mysqli_query($conn, "DELETE FROM products WHERE id = $id");
+  header("Location: product.php");
 ?>
